Add info on using `drop_caches` nondestructively
I was directed to linxuatemyram.com due to WSL2 memory usage and found it annoying and disappointing that it didn't actually provide any useful info. The tone of the paragraph above my PR isn't really the nicest when RAM usage is an issue, like in this case. At the very least this will lessen any feelings of frustration for future readers.